refactor: remove omitAndDeleteField action from adapter
