+++ lib/adapters/REST/endpoints/content-type.ts
@@ -82,52 +82,3 @@ export const publish = (
 export const unpublish = (http: AxiosInstance, params: GetContentTypeParams) => {
   return raw.del<ContentTypeProps>(http, getContentTypeUrl(params) + '/published')
 }
-
-type OmitOrDelete = 'omitted' | 'deleted'
-
-/**
- * @private
- * @param id - unique ID of the field
- * @param key - the attribute on the field to change
- * @param value - the value to set the attribute to
- */
-const findAndUpdateField = function (
-  contentType: ContentTypeProps,
-  fieldId: string,
-  omitOrDelete: OmitOrDelete
-) {
-  const field = contentType.fields.find((field) => field.id === fieldId)
-  if (!field) {
-    return Promise.reject(
-      new Error(
-        `Tried to omitAndDeleteField on a nonexistent field, ${fieldId}, on the content type ${contentType.name}.`
-      )
-    )
-  }
-  // @ts-expect-error
-  field[omitOrDelete] = true
-
-  return Promise.resolve(contentType)
-}
-
-export const omitAndDeleteField = (
-  http: AxiosInstance,
-  {
-    fieldId,
-    ...params
-  }: GetContentTypeParams & {
-    fieldId: string
-  },
-  contentType: ContentTypeProps
-) => {
-  return findAndUpdateField(contentType, fieldId, 'omitted')
-    .then((newContentType) => {
-      return update(http, params, newContentType)
-    })
-    .then((newContentType) => {
-      return findAndUpdateField(newContentType, fieldId, 'deleted')
-    })
-    .then((newContentType) => {
-      return update(http, params, newContentType)
-    })
-}
